test(button): restore to original structure for secondary variant screenshot tests
