+++ packages/core/src/components/button/test/secondary/button.e2e.ts-snapshots/tds-button-secondary-renders-secondary-button-correctly-1-linux.png
