--- packages/core/src/components/button/test/secondary/primary/darkmode/button.e2e.ts
@@ -1,18 +1,0 @@
-import { test } from 'stencil-playwright';
-import { expect } from '@playwright/test';
-
-const componentTestPath = 'src/components/button/test/secondary/primary/darkmode/index.html';
-
-test.describe.parallel('tds-button-secondary-primary-darkmode', () => {
-  test.beforeEach(async ({ page }) => {
-    await page.goto(componentTestPath);
-  });
-
-  test('renders secondary button correctly', async ({ page }) => {
-    const button = page.getByTestId('tds-button-testid');
-    await expect(button).toHaveCount(1);
-
-    /* Check diff on screenshot */
-    await expect(page).toHaveScreenshot({ maxDiffPixels: 0 });
-  });
-});
