--- packages/core/src/components/button/test/secondary/primary/lightmode/index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Button - Secondary</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0" />
-    <link rel="stylesheet" href="../../../../../../dist/tegel/tegel.css" />
-    <script type="module">
-      import { defineCustomElements } from '../../../../../../loader/index.es2017.js';
-      defineCustomElements();
-    </script>
-  </head>
-
-  <body style="padding: 20px">
-    <tds-button
-      text="Button"
-      variant="secondary"
-      mode-variant="primary"
-      size="md"
-      data-testid="tds-button-testid"
-    ></tds-button>
-  </body>
-</html>
